LayoutTests/fast/forms/select/menulist-focusable-items.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<select>
    <option tabindex="0" id="optionA">A</option>
    <optgroup tabindex="0" id="optgroupB" label="B">
        <option tabindex="0" id="optionC">C</option>
    </optgroup>
</select>
<option tabindex="0" id="optionD">D</option>
<script>
description('option/optgroup in menulist are never focusable even with a tabindex');
function isFocusable(id) {
    var element = document.getElementById(id);
    element.focus();
    return document.activeElement === element;
}
shouldBeFalse('isFocusable("optionA")');
shouldBeFalse('isFocusable("optgroupB")');
shouldBeFalse('isFocusable("optionC")');
shouldBeTrue('isFocusable("optionD")');
</script>
</body>
</html>
