third_party/blink/web_tests/platform/generic/fast/forms/select/menulist-focusable-items-expected.txt

option/optgroup in menulist are never focusable even with a tabindex

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isFocusable("optionA") is false
PASS isFocusable("optgroupB") is false
PASS isFocusable("optionC") is false
PASS isFocusable("optionD") is true
PASS successfullyParsed is true

TEST COMPLETE

A
C
D
